UI Overhaul
Semantic UI wasn't cutting it. This is a reversion to bootstrap, along
with a few small additions such as a footer and some placeholder text on
the homepage.

The minimum viable product is here! We have a reasonable workflow for
creating stories and chapters, we have extremely basic authentication,
we have chapter navigation. Life is good.